requirements.txt
httplib2==0.20.2
psycopg2-binary==2.9.3
pyparsing==3.0.7
git+https://github.com/gbvsilva/podio-py.git#egg=pypodio2
